backend/requirements.txt
uvicorn
fastapi
faiss-cpu
sentence-transformers
pdfplumber
numpy
langchain
python-dotenv
pydantic
